+++ test/end-to-end/fixtures/console-ability-changed/server.js
@@ -7,11 +7,12 @@ const domapic = require('../../../../index')
 domapic.createModule({
   packagePath: path.resolve(__dirname)
 }).then((service) => {
-  let lastCharacter
+  let lastCharacter = ''
 
-  const consoleLog = function (data) {
+  const consoleLog = function (data, eventName) {
     lastCharacter = data
     console.log(`Printing into console: ${data}`)
+    service.events.emit(eventName, data)
   }
 
   return service.register({
@@ -33,8 +34,40 @@ domapic.createModule({
         description: 'Print the received character into stdout',
         auth: false,
         handler: (data) => {
-          consoleLog(data)
+          consoleLog(data, 'stdout')
           return Promise.resolve(data)
+        }
+      }
+    },
+    consoleNoData: {
+      description: 'Handle console log with no data before, now has data',
+      data: {
+        type: 'number',
+        maximum: 10,
+        minimum: 1
+      },
+      event: {
+        description: 'Console with data received'
+      },
+      action: {
+        description: 'Print the received character into console',
+        handler: (data) => {
+          consoleLog(data, 'consoleNoData')
+          return Promise.resolve(data)
+        }
+      }
+    },
+    consoleNumeric: {
+      description: 'Handle console log with numeric data before, now has no data',
+      event: {
+        description: 'Console with numeric data received'
+      },
+      action: {
+        description: 'Print the received number into console',
+        handler: () => {
+          console.log('Printing consoleNumeric with no data')
+          service.events.emit('consoleNumeric')
+          return Promise.resolve()
         }
       }
     }