+++ test/end-to-end/specs/console-ability-changed.specs.js
@@ -1,5 +1,6 @@
 
 const test = require('narval')
+const testUtils = require('narval/utils')
 
 const utils = require('./utils')
 
@@ -38,7 +39,7 @@ test.describe('when connection with controller was successful', function () {
       .then(response => {
         const ability = response.body.find(ability => ability.name === 'stdout')
         return Promise.all([
-          test.expect(response.body.length).to.equal(1),
+          test.expect(response.body.length).to.equal(3),
           test.expect(ability._service).to.equal(serviceId),
           test.expect(ability._user).to.equal(serviceUserId),
           test.expect(ability.event).to.equal(false),
@@ -53,5 +54,152 @@ test.describe('when connection with controller was successful', function () {
           test.expect(ability.eventDescription).to.equal(undefined)
         ])
       })
+  })
+})
+
+test.describe('when using controller action api to dispatch ability that previously has no data', function () {
+  let consoleAbilityId
+  this.timeout(10000)
+  const controllerConnection = new utils.ControllerConnection()
+
+  test.before(() => {
+    return controllerConnection.request('/abilities')
+      .then(response => {
+        const ability = response.body.find(ability => ability.name === 'consoleNoData')
+        consoleAbilityId = ability._id
+        return Promise.resolve()
+      })
+  })
+
+  test.it('should reject the request if no data is received', () => {
+    return controllerConnection.request(`/abilities/${consoleAbilityId}/action`, {
+      method: 'POST'
+    }).then(response => {
+      return utils.waitOnestimatedStartTime(100)
+        .then(() => {
+          return testUtils.logs.combined('service')
+            .then((log) => {
+              return Promise.all([
+                test.expect(response.statusCode).to.equal(422)
+              ])
+            })
+        })
+    })
+  })
+
+  test.it('should print into console a log when right data is received', () => {
+    return controllerConnection.request(`/abilities/${consoleAbilityId}/action`, {
+      method: 'POST',
+      body: {
+        data: 5
+      }
+    }).then(response => {
+      return utils.waitOnestimatedStartTime(100)
+        .then(() => {
+          return testUtils.logs.combined('service')
+            .then((log) => {
+              return Promise.all([
+                test.expect(response.statusCode).to.equal(201),
+                test.expect(log).to.contain(`Printing into console: 5`)
+              ])
+            })
+        })
+    })
+  })
+
+  test.it('should trigger related event, and it should be saved into controller database', () => {
+    return controllerConnection.request(`/abilities/${consoleAbilityId}/action`, {
+      method: 'POST',
+      body: {
+        data: 6
+      }
+    }).then(response => {
+      return utils.waitOnestimatedStartTime(200)
+        .then(() => {
+          return controllerConnection.request(`/logs`, {
+            method: 'GET'
+          }).then(logsResponse => {
+            const actionLog = logsResponse.body.find(log => log._ability === consoleAbilityId && log.data === 6 && log.type === 'action')
+            const eventLog = logsResponse.body.find(log => log._ability === consoleAbilityId && log.data === 6 && log.type === 'event')
+            return Promise.all([
+              test.expect(logsResponse.statusCode).to.equal(200),
+              test.expect(actionLog).to.not.be.undefined(),
+              test.expect(eventLog).to.not.be.undefined()
+            ])
+          })
+        })
+    })
+  })
+})
+
+test.describe('when using controller action api to dispatch ability action with numeric data before, but now has not data', function () {
+  let consoleAbilityId
+  this.timeout(10000)
+  const controllerConnection = new utils.ControllerConnection()
+
+  test.before(() => {
+    return controllerConnection.request('/abilities')
+      .then(response => {
+        const ability = response.body.find(ability => ability.name === 'consoleNumeric')
+        consoleAbilityId = ability._id
+        return Promise.resolve()
+      })
+  })
+
+  test.it('should reject the request if data is invalid', () => {
+    return controllerConnection.request(`/abilities/${consoleAbilityId}/action`, {
+      method: 'POST',
+      body: {
+        data: 'foo'
+      }
+    }).then(response => {
+      return utils.waitOnestimatedStartTime(100)
+        .then(() => {
+          return testUtils.logs.combined('service')
+            .then((log) => {
+              return Promise.all([
+                test.expect(response.statusCode).to.equal(422)
+              ])
+            })
+        })
+    })
+  })
+
+  test.it('should print into console a log when no data is received', () => {
+    return controllerConnection.request(`/abilities/${consoleAbilityId}/action`, {
+      method: 'POST'
+    }).then(response => {
+      return utils.waitOnestimatedStartTime(100)
+        .then(() => {
+          return testUtils.logs.combined('service')
+            .then((log) => {
+              return Promise.all([
+                test.expect(response.statusCode).to.equal(201),
+                test.expect(log).to.contain(`Printing consoleNumeric with no data`)
+              ])
+            })
+        })
+    })
+  })
+
+  test.it('should trigger related event, and it should be saved into controller database', () => {
+    return controllerConnection.request(`/abilities/${consoleAbilityId}/action`, {
+      method: 'POST'
+    }).then(response => {
+      return utils.waitOnestimatedStartTime(200)
+        .then(() => {
+          return controllerConnection.request(`/logs`, {
+            method: 'GET'
+          }).then(logsResponse => {
+            const actionLog = logsResponse.body.find(log => log._ability === consoleAbilityId && log.type === 'action')
+            const eventLog = logsResponse.body.find(log => log._ability === consoleAbilityId && log.type === 'event')
+            return Promise.all([
+              test.expect(logsResponse.statusCode).to.equal(200),
+              test.expect(actionLog).to.not.be.undefined(),
+              test.expect(eventLog).to.not.be.undefined()
+            ])
+          })
+        })
+    })
   })
 })