+++ test/functional/fixtures/abilities-wrong-responses/server.js
@@ -7,14 +7,8 @@ const domapic = require('../../../../index')
 domapic.createModule({
   packagePath: path.resolve(__dirname)
 }).then((service) => {
-  let lastCharacter
-
   const consoleLog = function (data, ability) {
-    lastCharacter = data
     service.tracer.info('Console Called:', data)
-      .then(() => {
-        service.events.emit(ability, data)
-      })
   }
 
   return service.register({
@@ -29,14 +23,14 @@ domapic.createModule({
       state: {
         description: 'Last boolean printed in console',
         handler: () => {
-          return Promise.resolve(lastCharacter)
+          return Promise.resolve(null)
         }
       },
       action: {
         description: 'Print the received boolean into console',
         handler: (data) => {
           consoleLog(data, 'booleanConsole')
-          return Promise.resolve(data)
+          return Promise.resolve(3)
         }
       }
     },
@@ -55,14 +49,14 @@ domapic.createModule({
       state: {
         description: 'Last email printed in console',
         handler: () => {
-          return Promise.resolve(lastCharacter)
+          return Promise.resolve('foo')
         }
       },
       action: {
         description: 'Print the received email into console',
         handler: (data) => {
           consoleLog(data, 'emailConsole')
-          return Promise.resolve(data)
+          return Promise.resolve(true)
         }
       }
     },
@@ -78,14 +72,14 @@ domapic.createModule({
       state: {
         description: 'Last string from enum printed in console',
         handler: () => {
-          return Promise.resolve(lastCharacter)
+          return Promise.resolve(false)
         }
       },
       action: {
         description: 'Print the received string from enum into console',
         handler: (data) => {
           consoleLog(data, 'enumConsole')
-          return Promise.resolve(data)
+          return Promise.resolve('foo3')
         }
       }
     },
@@ -103,14 +97,14 @@ domapic.createModule({
       state: {
         description: 'Last number printed in console',
         handler: () => {
-          return Promise.resolve(lastCharacter)
+          return Promise.resolve('foo')
         }
       },
       action: {
         description: 'Print the received number into console',
         handler: (data) => {
           consoleLog(data, 'numericConsole')
-          return Promise.resolve(data)
+          return Promise.resolve(150)
         }
       }
     },
@@ -130,14 +124,14 @@ domapic.createModule({
       state: {
         description: 'Last number printed in console',
         handler: () => {
-          return Promise.resolve(lastCharacter)
+          return Promise.resolve(4)
         }
       },
       action: {
         description: 'Print the received number from enum into console',
         handler: (data) => {
           consoleLog(data, 'numericEnumConsole')
-          return Promise.resolve(data)
+          return Promise.resolve(false)
         }
       }
     },
@@ -149,21 +143,8 @@ domapic.createModule({
       action: {
         description: 'Print hello into console',
         handler: () => {
-          consoleLog(undefined, 'noDataConsole')
-          return Promise.resolve()
-        }
-      }
-    },
-    noDataWrongConsole: {
-      description: 'Prints hello into console',
-      event: {
-        description: 'Console has just printed hello'
-      },
-      action: {
-        description: 'Print hello into console',
-        handler: () => {
-          consoleLog('hello', 'noDataWrongConsole')
-          return Promise.resolve()
+          consoleLog('hello', 'noDataConsole')
+          return Promise.resolve(5)
         }
       }
     }