+++ test/functional/plugin-specs/config-api.specs.js
@@ -28,7 +28,8 @@ test.describe('config api', function () {
           exampleOption: false,
           authDisabled: [],
           hostName: utils.SERVICE_HOST,
-          path: path.resolve(__dirname, '..', '..', '..', utils.DOMAPIC_PATH)
+          path: path.resolve(__dirname, '..', '..', '..', utils.DOMAPIC_PATH),
+          rejectUntrusted: false
         })
       ])
     })