+++ test/functional/specs/storage.specs.js
@@ -1,0 +1,12 @@
+const test = require('narval')
+
+const utils = require('./utils')
+
+test.describe('service storage', function () {
+  test.it('should have saved test value to storage', () => {
+    return utils.readStorage()
+      .then(storedData => {
+        return test.expect(storedData.storageTest).to.equal('foo-storage-value')
+      })
+  })
+})