+++ CHANGELOG.md
@@ -10,6 +10,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [1.0.0-alpha.3] - 2018-12-01
+### Added
+- Expose storage method to services
+
+### Changed
+- Upgrade domapic-base version, which fixes a problem in Client concurrent requests.
+- Do not demand data in abilities. Still mandatory in abilities with state defined.
+
+### Fixed
+- Extend exposed cli options with service options
+
 ## [1.0.0-alpha.2] - 2018-11-20
 ### BREAKING CHANGES
 - Changed Controller uris to adapt them to Controller version 1.0.0-alpha.9